agent-factory/skill-guardrails/requirements.txt
# Skill Guardrails - Python Dependencies

# Core
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0

# HTTP Client for Gemini API
httpx>=0.26.0
tenacity>=8.2.0

# Pattern Matching
regex>=2023.12.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.0

# Hashing & Signatures
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Reporting
jinja2>=3.1.0

# Optional: Docker SDK for sandbox (Level 3)
# docker>=7.0.0
